refactoring

+++ backend-gitlab/src/api/accept_merge_request.rs
@@ -2,7 +2,7 @@ pub async fn accept_merge_request(
     host: &str,
     token: &str,
     project_id: usize,
-    merge_request_iid: i32,
+    merge_request_iid: u64,
     request: &crate::models::accept_merge_request::Request,
 ) -> Result<crate::models::accept_merge_request::Response, reqwest::Error> {
     let url = format!(
